.py + .kv combined to colorusage.py and placed in examples/widgets/colorusage.py
